+++ sdk/azcore/runtime/pager_test.go
@@ -264,52 +264,59 @@ func TestFetcherForNextLink(t *testing.T) {
 	pl := exported.NewPipeline(srv)
 
 	srv.AppendResponse()
-	createReqCalled := false
+	firstReqCalled := false
 	resp, err := FetcherForNextLink(context.Background(), pl, "", func(ctx context.Context) (*policy.Request, error) {
-		createReqCalled = true
+		firstReqCalled = true
 		return NewRequest(ctx, http.MethodGet, srv.URL())
-	})
+	}, nil)
 	require.NoError(t, err)
-	require.True(t, createReqCalled)
+	require.True(t, firstReqCalled)
 	require.NotNil(t, resp)
 	require.EqualValues(t, http.StatusOK, resp.StatusCode)
 
 	srv.AppendResponse()
-	createReqCalled = false
+	firstReqCalled = false
+	nextReqCalled := false
 	resp, err = FetcherForNextLink(context.Background(), pl, srv.URL(), func(ctx context.Context) (*policy.Request, error) {
-		createReqCalled = true
+		firstReqCalled = true
 		return NewRequest(ctx, http.MethodGet, srv.URL())
+	}, &FetcherForNextLinkOptions{
+		NextReq: func(ctx context.Context, s string) (*policy.Request, error) {
+			nextReqCalled = true
+			return NewRequest(ctx, http.MethodGet, srv.URL())
+		},
 	})
 	require.NoError(t, err)
-	require.False(t, createReqCalled)
+	require.False(t, firstReqCalled)
+	require.True(t, nextReqCalled)
 	require.NotNil(t, resp)
 	require.EqualValues(t, http.StatusOK, resp.StatusCode)
 
 	resp, err = FetcherForNextLink(context.Background(), pl, "", func(ctx context.Context) (*policy.Request, error) {
 		return nil, errors.New("failed")
-	})
+	}, &FetcherForNextLinkOptions{})
 	require.Error(t, err)
 	require.Nil(t, resp)
 
 	srv.AppendError(errors.New("failed"))
 	resp, err = FetcherForNextLink(context.Background(), pl, "", func(ctx context.Context) (*policy.Request, error) {
-		createReqCalled = true
+		firstReqCalled = true
 		return NewRequest(ctx, http.MethodGet, srv.URL())
-	})
+	}, &FetcherForNextLinkOptions{})
 	require.Error(t, err)
-	require.True(t, createReqCalled)
+	require.True(t, firstReqCalled)
 	require.Nil(t, resp)
 
 	srv.AppendResponse(mock.WithStatusCode(http.StatusBadRequest), mock.WithBody([]byte(`{ "error": { "code": "InvalidResource", "message": "doesn't exist" } }`)))
-	createReqCalled = false
+	firstReqCalled = false
 	resp, err = FetcherForNextLink(context.Background(), pl, srv.URL(), func(ctx context.Context) (*policy.Request, error) {
-		createReqCalled = true
+		firstReqCalled = true
 		return NewRequest(ctx, http.MethodGet, srv.URL())
-	})
+	}, nil)
 	require.Error(t, err)
 	var respErr *exported.ResponseError
 	require.ErrorAs(t, err, &respErr)
 	require.EqualValues(t, "InvalidResource", respErr.ErrorCode)
-	require.False(t, createReqCalled)
+	require.False(t, firstReqCalled)
 	require.Nil(t, resp)
 }